insures that user is authenticated and if is admin (cannot edit, post, delete even if try to get it by URL).

+++ controllers/articles.js
@@ -2,6 +2,17 @@ const Articles = require('mongoose').model('Articles');
 
 module.exports = {
     articlesGet: (req, res) => {
+        let id = req.params.id;
+        if(!req.isAuthenticated()){
+
+            res.redirect('/users/login');
+            return;
+        }
+        if(req.isAuthenticated() && req.user.email !== "admin@admin.com"){
+            res.redirect('/');
+            return;
+        }
+
         res.render('articles/create');
     },
     articlesPost: (req, res) => {
@@ -77,7 +88,17 @@ module.exports = {
 
     articlesEditGet: (req, res) => {
         let id = req.params.id;
+        if(!req.isAuthenticated()){
+            let returnUrl = `articles/edit/${id}`;
+            req.session.returnUlr = returnUrl;
 
+            res.redirect('/users/login');
+            return;
+        }
+        if(req.isAuthenticated() && req.user.email !== "admin@admin.com"){
+            res.redirect('/');
+            return;
+        }
         Articles.findById(id).populate('author').then(article => {
             res.render('articles/edit', article)
         })
@@ -85,6 +106,15 @@ module.exports = {
     },
     articlesEditPost: (req, res) => {
         let body = req.body;
+        if(!req.isAuthenticated()){
+
+            res.redirect('/users/login');
+            return;
+        }
+        if(req.isAuthenticated() && req.user.email !== "admin@admin.com"){
+            res.redirect('/');
+            return;
+        }
         Articles.findOneAndUpdate({
             id: body.id
         }, body, {
@@ -96,6 +126,17 @@ module.exports = {
     articlesDeletePost: (req, res) => {
         let id = req.params.id;
 
+        if(!req.isAuthenticated()){
+            let returnUrl = `articles/delete/${id}`;
+            req.session.returnUlr = returnUrl;
+
+            res.redirect('/users/login');
+            return;
+        }
+        if(req.isAuthenticated() && req.user.email !== "admin@admin.com"){
+            res.redirect('/');
+            return;
+        }
         Articles.findOneAndRemove({_id: id}).then(article => {
             let author = article.author;
